function/pyproject.toml
[project]
name = "function"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "aws-lambda-powertools[parser]>=3.5.0",
    "boto3>=1.36.16",
    "types-boto3[essential]>=1.36.16",
]
